requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
Pillow==10.4.0
tensorflow==2.15.0
google-generativeai==0.7.2
python-decouple==3.8
drf-spectacular==0.27.2
gunicorn==22.0.0
whitenoise==6.7.0
numpy
